Notification refactor singletons refactor and pass components
