+++ PasswordMenager/frontend/src/components/AppBar/AccountContainer/AccountInfo/AccountInfoView/NotificationCardView/NotificationModals/ModalsTexts/index.tsx
@@ -1,0 +1,3 @@
+export const ActiveModal = <div>Active notification?</div>;
+export const SuspendModal = <div>Suspend notification?</div>;
+export const DeleteModal = <div>Are you sure to delete notification?</div>;
