+++ PasswordMenager/frontend/src/components/AppBar/AccountContainer/AccountInfo/AccountInfoView/NotificationCardView/NotificationModals/mappers.ts
@@ -1,0 +1,19 @@
+import {
+  ActivateNotification,
+  DeleteNotification,
+  SuspendNotifcation,
+} from "../utils";
+import { ActiveModal, DeleteModal, SuspendModal } from "./ModalsTexts";
+
+export const ComponentMapper = {
+  activate: ActiveModal,
+  suspend: SuspendModal,
+  delete: DeleteModal,
+  "": undefined,
+};
+export const ActionHandlerMapper = {
+  activate: ActivateNotification,
+  suspend: SuspendNotifcation,
+  delete: DeleteNotification,
+  "": null,
+};
