+++ PasswordMenager/frontend/src/components/PassComponent/EmptyEntries/index.tsx
@@ -1,0 +1,10 @@
+import { Translation } from "../../Translation";
+import { EmptyEntries, Text } from "../component.styled";
+
+export const EmptyEntriesComponent = () => {
+  return (
+    <EmptyEntries>
+      <Text>{Translation("entries.noEntriesAvailable")}</Text>
+    </EmptyEntries>
+  );
+};
